+++ src/services/direct/index.ts
@@ -3,7 +3,9 @@ export * from './account-mapping'
 export * from './token-service'
 export * from './common'
 export * from './health'
+export * from './fireblocks-config'
 export * from './fireblocks-provider'
+export * from './dfns-config'
 export * from './dfns-provider'
 export * from './omnibus-delegate'
 export * from './mapping-validator'
