+++ tests/utils/fireblocks-test-environment.ts
@@ -20,7 +20,7 @@ import createApp from "../../src/app";
 import {
   createFireblocksEthersProvider,
   FireblocksAppConfig,
-} from "../../src/config";
+} from "../../src/services/direct/fireblocks-config";
 import { randomPort } from "./utils";
 
 dotenv.config({ path: resolve(process.cwd(), ".env.fireblocks") });
